Address PR #152 review comments
- Move CRON_SECRET, cron_request_source() and is_cron_request() from
  firebase.py to basics.py, their natural home. firebase.py's own
  /connect/update handler imports is_cron_request lazily, since a
  module-level import would be circular (basics -> skrafluser ->
  firebase) - which is why these helpers had historically ended up in
  firebase.py in the first place. skraflstats.py and web.py now import
  from basics; firebase.py drops its now-unused os/running_local/
  ON_GAE/ON_GCP imports.

- Rename GoogleSecretManager to GoogleSecretProvider for naming
  consistency with SecretProvider/EnvSecretProvider, and remove the
  unused SecretManager compatibility alias (no importers remained).

- db/postgresql/__init__: drop the redundant "import X as X" (the
  __all__ declaration already marks the re-export) and cache the
  lazily imported PostgreSQLBackend in the module namespace so
  subsequent attribute accesses bypass __getattr__.

The session.flush() convention question (ChatRepository.delete_for_user)
is answered on the PR: it follows the established pattern used by every
other bulk delete in the module.

Verified: tests/api_e2e 125 passed, test/ 49 passed, pyright clean,
env-provider isolation smoke test passes.
